Refactor: Mocks
