--- __mocks__/build-subject.ts
@@ -1,8 +1,0 @@
-import { loadObservers } from "../lib/src";
-import { notifySalesDepartmentObservers } from "./notify-sales-observers";
-import { sendEmailsObservers } from "./send-email-observers";
-
-export const observers = loadObservers([
-  ...sendEmailsObservers,
-  ...notifySalesDepartmentObservers,
-]);
