+++ __mocks__/observers-container/loaded-observers.ts
@@ -1,0 +1,8 @@
+import { loadObservers } from "../../lib/src";
+import { notifySalesDepartmentObservers } from "./observers/notify-sales-observers";
+import { sendEmailsObservers } from "./observers/send-email-observers";
+
+export const observers = loadObservers([
+  ...sendEmailsObservers,
+  ...notifySalesDepartmentObservers,
+]);
